refactor: split OAuth pool health checks
